email/requirements.txt
python-dotenv
pillow
openpyxl
requests
pywin32; platform_system == "Windows"
python-telegram-bot
